Move EnipCodec to enip-client module
